app/catalog-service/requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-logging==3.9.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
